使代码适配Ubuntu，同时封装了代码，增强可读性
