+++ README.md
@@ -1,8 +1,13 @@
 <h2>如何运行</h2><br>
-把 **src/Experi** 里面的代码贴到main.cpp上，再在项目主文件夹终端输入```make run```即可，或者编译成功后运行output文件下的main.exe可执行文件。<br>
+确保ubuntu版本安装以下库
+```<br>
+sudo apt update<br>
+sudo apt install mesa-utils build-essential libglfw3 libglfw3-dev<br>
+```<br>
+把 **Experi** 里面相应实验的代码贴到main.cpp上，再在项目主文件夹终端输入```make run```即可，或者编译成功后运行output文件下的main.exe可执行文件。<br>
 如果输入```make run```后提示没有make命令，改为输入```mingw32-make run```。<br>
 
-<h2>one.cpp</h2> <br>
+<h2>one.txt</h2> <br>
 绘制直线和圆弧，窗口启动之后，可以按键盘上的l或者c来切换直线或者圆弧模式，按```ctrl+z```可以撤销上一次操作，按```ctrl+s```可以将图像保存到当前目录下(output.tga)<br>
 直线模式：鼠标左键设置起始点，此时直线另一端会随鼠标移动，再次按下左键完成直线的绘制<br>
 圆弧模式：鼠标左键设置圆心，再次按下左键固定半径，此时会以顺时针角度绘制圆弧，再次按下左键完成绘制<br>
